+++ src/modules/searchBar/components/SearchBar.vue
@@ -148,11 +148,9 @@ export default {
         currentComponentSide: {
             handler (newVal) {
                 if (newVal === "root" || (newVal === "layerSelection" && this.addLayerButtonSearchActive === true)) {
-                    this.removePolygonMarker();
-                    this.removePointMarker();
                     this.$refs?.searchInput.blur();
                     if (newVal === "root") {
-                        this.searchInputValue = "";
+                        this.setSearchResultsActive(false);
                         this.navigateBack(this.currentSide);
                         if (this.side) {
                             this.switchToRoot(this.side);
@@ -172,8 +170,14 @@ export default {
         * Watcher to check the searchInputValue for layerSelection module
         */
         searchInputValue: {
-            handler () {
-                this.checkCurrentComponent(this.currentComponentSide);
+            handler (value) {
+                if (value === "") {
+                    this.removePointMarker();
+                    this.removePolygonMarker();
+                }
+                else {
+                    this.checkCurrentComponent(this.currentComponentSide);
+                }
             },
             deep: true
         }